pyproject.toml
[tool.poetry]
name = "webvr-project"
version = "0.1.0"
description = ""
authors = ["babuuh <kiharajoseph72@gamil.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0"
suit = "^2.0.2"
django-mysql = "^4.3.0"
django-crispy-forms = "^1.13.0"
django-extra-views = "^0.14.0"
django-colorful = "^1.3"
Pillow = "^8.4.0"
django-filter = "^21.1"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
